feature(ui-ux): add/remove collateralV2, address pr comments
