+++ mobile-app/app/components/BottomSheetTokenList.tsx
@@ -20,8 +20,8 @@ import {
   ThemedFlatListV2,
   ThemedTextV2,
   ThemedTouchableOpacityV2,
-  ThemedViewV2,
 } from "./themed";
+import { CollateralFactorTag } from "./CollateralFactorTag";
 import { SymbolIcon } from "./SymbolIcon";
 
 interface BottomSheetTokenListProps {
@@ -219,25 +219,3 @@ export const BottomSheetTokenList = ({
       />
     );
   });
-
-export function CollateralFactorTag({
-  factor,
-}: {
-  factor?: string;
-}): JSX.Element | null {
-  const DEFAULT_FACTOR = "1";
-  return factor !== undefined && factor !== DEFAULT_FACTOR ? (
-    <ThemedViewV2
-      style={tailwind("h-5 flex flex-row items-center px-2 rounded border-0.5")}
-      light={tailwind("border-mono-light-v2-700")}
-      dark={tailwind("border-mono-dark-v2-700")}
-    >
-      <ThemedTextV2
-        light={tailwind("text-xs font-semibold-v2 text-mono-light-v2-700")}
-        dark={tailwind("text-xs font-semibold-v2 text-mono-dark-v2-700")}
-      >
-        {`${factor}x`}
-      </ThemedTextV2>
-    </ThemedViewV2>
-  ) : null;
-}
