+++ mobile-app/app/styles.json
@@ -3657,9 +3657,6 @@
 	"mb-5": {
 		"marginBottom": 20
 	},
-	"mb-5.5": {
-		"marginBottom": 22
-	},
 	"mb-6": {
 		"marginBottom": 24
 	},
